pyproject.toml
[project]
name = "powerpoint"
version = "0.1.0"
description = "A MCP server project that creates powerpoint presentations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "mcp>=1.3.0",
 "openai>=1.77.0",
 "pillow>=11.1.0",
 "python-pptx>=1.0.2",
 "requests>=2.32.3",
 "together>=1.4.1",
]
[[project.authors]]
name = "Russell Ashby"
email = "158164671+supercurses@users.noreply.github.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
powerpoint = "powerpoint:main"
